requirements.txt
altgraph==0.17.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
browser-cookie3==0.19.1
browsercookie==0.7.9
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gspread==6.0.2
httplib2==0.22.0
idna==3.6
importlib_metadata==8.0.0
ipython==8.12.3
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
customtkinter==5.2.2
darkdetect==0.8.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyboard==0.13.5
keyring==25.2.1
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.20.0
pycryptodomex==3.20.0
PyDrive==1.3.1
pyglet==2.0.15
Pygments==2.18.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pykeyboard==0.1.5
pyparsing==3.1.2
Pyrogram==2.0.106
pyscard==2.0.8
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pyusb==1.2.1
plyer==2.1.0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.4.0
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
StrEnum==0.4.15
TgCrypto==1.2.5
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.19.2
